+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Dropped all support for python 2.7
 - Swapped the mutable and immutable defaults
 - Removed the inclusion of the constants at the top level
+- Removed all mutable timestamps
 
 ## 1.7.2
 - This is the final version of this library to support Python 2.7